HEEDLS-495 Refactor into SetPasswordControllerBase; variable renaming
